# microservices/core-services/order-service/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Order Service API
  description: |
    Order Management Service API for the Inventory Management System.
    This API provides endpoints for managing orders, including creation, status updates,
    cancellation, and retrieval of order information.
  version: 3.0.0
  contact:
    name: API Support
    email: support@inventory-system.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8082
    description: Development server
  - url: https://api.order-service.inventory-system.com
    description: Production server
...
paths:
...
  /api/orders/{orderId}:
    get:
      tags:
        - Orders
      summary: Get order by ID
      description: Retrieve a specific order by its ID
      operationId: getOrderById
      parameters:
        - name: orderId
          in: path
          required: true
          description: Order ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    $ref: '#/components/schemas/OrderDTO'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    put:
      tags:
        - Orders
      summary: Update order
      description: Update an existing order's information
      operationId: updateOrder
      parameters:
        - name: orderId
          in: path
          required: true
          description: Order ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdateRequest'
      responses:
        '200':
          description: Order updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Order not found

    delete:
      tags:
        - Orders
      summary: Delete order
      description: Delete an order by ID (only PENDING orders can be deleted)
      operationId: deleteOrder
      parameters:
        - name: orderId
          in: path
          required: true
          description: Order ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Order deleted successfully
        '400':
          description: Cannot delete non-PENDING order
        '404':
          description: Order not found
  /api/orders/{orderId}/status:
    put:
      tags:
        - Orders
      summary: Update order status
      description: |
        Update the status of an order.
        Valid transitions:
        - PENDING -> PROCESSING, CANCELLED
        - PROCESSING -> SHIPPED, CANCELLED
        - SHIPPED -> DELIVERED
        - DELIVERED -> REFUNDING
        - REFUNDING -> REFUNDED, COMPLETED
      operationId: updateOrderStatus
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [PENDING, PROCESSING, COMPLETED, CANCELLED, REFUNDING, REFUNDED]
      responses:
        '200':
          description: Status updated successfully
        '400':
          description: Invalid status transition
...
  /api/orders/number/{orderNumber}:
    get:
      tags:
        - Orders
      summary: Get order by order number
      operationId: getOrderByNumber
      parameters:
        - name: orderNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDTO'
        '404':
          description: Order not found
...
components:
  schemas:
    OrderDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderNumber:
          type: string
          example: "ORD-2024-00001"
        customerId:
          type: integer
          format: int64
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        shippingAddress:
          type: string
        billingAddress:
          type: string
        status:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, CANCELLED, REFUNDING, REFUNDED]
        paymentMethod:
          type: string
          enum: [CREDIT_CARD, DEBIT_CARD, PAYPAL, BANK_TRANSFER, CASH]
        paymentStatus:
          type: string
          enum: [PENDING, PAID, FAILED, REFUNDED]
        subtotal:
          type: number
          format: double
        taxAmount:
          type: number
          format: double
        shippingCost:
          type: number
          format: double
        discountAmount:
          type: number
          format: double
        totalAmount:
          type: number
          format: double
        itemCount:
          type: integer
        notes:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemDTO'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        shippedAt:
          type: string
          format: date-time
        deliveredAt:
          type: string
          format: date-time
    OrderItemDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        productName:
          type: string
        productImage:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
          format: double
        discount:
          type: number
          format: double
        totalPrice:
          type: number
          format: double
...
    OrderUpdateRequest:
      type: object
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        shippingAddress:
          type: string
        billingAddress:
          type: string
        notes:
          type: string
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: object
        timestamp:
          type: string
          format: date-time
...
security:
  - bearerAuth: []